Update patch set 1
Patch Set 1:

Brad,

may I ask you to handle the patch? I am not expert in Go, and the contributing process is quite frustrating. It is unlikely that I will contribute more code to Go, so it is quite pointless for both of us spending time on teaching me. The patch carries the idea on what needs to be done, can you please modify it so that it satisfies Go project standards and commit it?

P.S. I promise to handle your patch if you ever want to contribute to FreeBSD project. Just send it to me by email :)

Patch-set: 1
Reviewer: Gerrit User 17785 <17785@62eb7196-b449-3ce5-99f1-c037f21e1705>