Update patch set 3
Patch Set 3:

Ping? Even if we put x/net/idna in x/text/idna, this CL is independent of this. It would be good to not wait too long to put this in 1.9.

As core already includes the norm and width tables, unlike with x/net, this change doesn't impact the table sizes by much (width is reduced, the idna tables are added, bidi is optionally only needed for compilation).

Patch-set: 3
Reviewer: Gerrit User 5182 <5182@62eb7196-b449-3ce5-99f1-c037f21e1705>