cmd/compile: explicitly track whether Type.Width has been calculated
Prior to this CL, Type.Width != 0 was the mark
of a Type whose Width had been calculated.
As a result, dowidth always recalculated
the width of struct{}.
This, combined with the prohibition on calculating
the width of a FuncArgsStruct and the use of
struct{} as a function argument,
meant that there were circumstances in which 
it was forbidden to call dowidth on a type.
This inhibits refactoring to call dowidth automatically,
rather than explicitly.
Instead, track explicitly whether a struct's
width has been calculated.

Change-Id: Ie9a9fb5d924e7a2010c1904ae5e38ed4a38eaeb2