database/sql: allow drivers to support custom arg types
Previously all arguments were passed through driver.IsValid.
This checked arguments against a few fundamental go types and
prevented others from being passed in as arguments.

The new interface driver.NamedValueChecker may be defined
on both driver.Stmt and driver.Conn. This allows
this new interface to completely supersede the
driver.ColumnConverter interface as it can be used for
checking known statments and arbitrary query arguments.
The NamedValueChecker may be skipped with ErrSkip and
the default ColumnConverter used for all arguments,
or an individual argument may be omitted from
begin passed to the query at all. This is useful to pass
query specific options to the driver.

This interface remains opt-in. While drivers will need to
re-implement about 20 existing type checking and type
converting lines of code, this is dwarfed by the amount of time,
energy, and code that has been written to bypass these
checks.

Add a canonical Out argument wrapper to be passed
to OUTPUT parameters. This will unify checks that need to
be written in the NameValueChecker.

The statment number check is also moved to the argument
converter so the NamedValueChecker may remove arguments
passed to the query.

Fixes #13567
Fixes #18079
Updates #18417
Updates #17834
Updates #16235

Change-Id: I89088bd9cca4596a48bba37bfd20d987453ef237