cmd/compile: use MOVL instead of MOVQ for small constants on amd64
DO NOT SUBMIT
- needs benchmarks
(but should not effect performance negatively
execept with loop alignment changes,
and should have a tiny but overall positive
effect on instruction caches)

The encoding of MOVL to a register is 2 bytes shorter than for MOVQ.
The upper 32bit are automatically zeroed when MOVL to a register is used.

Replaces 1161 MOVQ by MOVL in the go binary.
As an example the mapaccess1_fast32 function shrinks by 4 byte.

Change-Id: I32a099b65f9e6d4ad7288ed48546655c534757d8