cmd/vet: check for missing error assignments
If a variable called "err" is declared in the outer scope, code
like this will compile without error:

if f(); err != nil {
	...
}

however we have seen that this is almost always a mistake, and
instead the code should be

if err := f(); err != nil {
	...
}

False positives seem rare. It's possible, though unusual, to make
f() a closure over the err variable.

This check does not attempt to detect if f() returns an error. It
simply looks for a variable called "err" that is being compared
to nil in a if statement, where the init statement was a function
call.

Change-Id: Ibf94bdb98322c2bfcb04bf369917dc0f7d3ff8c8