cmd/compile: simplify funcsyms
Every caller of makefuncsym is protected
by Ctxt.Flag_dynlink, and the !Ctxt.Flag_dynlink
check in funcsymname appears to be only to
compensate for that.

Sym.Fsym is used only to avoid adding duplicate
entries to funcsyms, but that is easily
accomplished by detecting the first lookup
vs subsequent lookups of the func sym name.

This avoids creating an unnecessary ONAME node
during funcsym, which eliminates a dependency
in the backend on Curfn and lineno.

It also makes the code a lot simpler and clearer.

Updates #15756

Does not pass toolstash-check,
but every instance I have investigated has been
trivial re-ordering of loads from the stack.

For example, before:

	0x03ef 01007 (/Users/josh/go/tip/src/crypto/rsa/rsa.go:118)	CALL	crypto.Hash.New(SB)
	0x03f4 01012 (/Users/josh/go/tip/src/crypto/rsa/rsa.go:118)	MOVQ	16(SP), AX
	0x03f9 01017 (/Users/josh/go/tip/src/crypto/rsa/rsa.go:118)	MOVQ	8(SP), CX

After:

	0x03ef 01007 (/Users/josh/go/tip/src/crypto/rsa/rsa.go:118)	CALL	crypto.Hash.New(SB)
	0x03f4 01012 (/Users/josh/go/tip/src/crypto/rsa/rsa.go:118)	MOVQ	8(SP), AX
	0x03f9 01017 (/Users/josh/go/tip/src/crypto/rsa/rsa.go:118)	MOVQ	16(SP), CX

In the before case, 16(SP) gets loaded before 8(SP).
It is reversed in the after case, with downstream
registers adjusted appropriately.

I am not sure why this CL triggers this slight
behavior change.

Change-Id: I7116c78fa39d975b7bd2c65a1d228749cf0dd46b