Update patch set 2
Patch Set 2:

So the race was between this n.Type assignment and the dowidth(fn.Type) call in (*state).call?

I think the proper fix then is in walkexpr's case ONAME. If n is potentially accessible outside the function (PEXTERN, PFUNC, PPARAM, or PPARAMOUT), it's not safe to modify at all. That SetAddable call isn't safe, nor is updateHasCall.

Patch-set: 2