syscall: add support for CLONE_NEWPID
Currently, the CLONE_NEWPID flag, if set,
is not considered by the Go exec package
when the process is created. Instead it
is directly passed to the child, which
executes it without question as part of the unshare
calls.

This leads to all sorts of problems for the Go runtime,
as described in issue 19716:
o When CLONE_NEWPID is set via unshare(2), the
  Go runtime only gets to fork once,
  leading to immediate failure
o the process name space is a bit weird; should the child
  mount /proc, it sees different
  PIDs before and after the mount. Hence, e.g., ps
  will give different output before and after the first
  /proc mount (but not the next <n> mounts of /proc)
o Should the child mount /proc without unsharing the
  mount namespace, /proc will instantly be broken for
  all processes outside the PID namespace. You can
  try this yourself as documented in the issue.

This change modifies exec for linux as follows:
o before the clone:
  if CLONE_NEWPID is set in Unshareflags, set it
  in Cloneflags and clear it in Unshareflags.
  This still gives the child a private PID namespace.
  But it will allow the child to fork more than once
  and still be in a private PID space.

  after the clone:
o Since CLONE_NEWS is set, the namespace is privatized.
o if CLONE_NEWPID is set in Cloneflags,
  mount /proc into the private name space. That way,
  the first pid is always seen as PID 1.
  This last change may be a bit controversial, as it assumes
  all users of private PID spaces want to see the first
  process as PID 1. Bbut it
  does ensure uniform behavior any time CLONE_NEWPID
  is set; further, users don't have to remember to
  mount /proc in the child to get the benefits of
  private PID namespaces. Finally, probably the most
  common use of private PID namespaces is for starting
  a /init, so this is a sensible thing to do.

The added tests check that CLONE_NEWPID features are
enabled via a helper test. If they are enabled, it kicks
off a helper and verifies that the helper thinks it is
PID 1.

This change is also tested on the u-root unshare command and
resolves our problems with the exec package.

Fixes #19716

Change-Id: I56b14d2f1a1e55e79dfd622f05eb7fa48f054364
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>