cmd/internal/obj/s390x: fix morestack cutoff for medium-sized stack frames
There is always 256 bytes available below the stackguard. Allow functions
with medium-sized stack frames to use this, potentially allowing them to
avoid growing the stack.

This CL changes the cutoff to match both the comment above it and the x86
code. I can't see any reason for the difference, hence this change.

Change-Id: I2afb1a7c686ae5a933e50903b31ea4106e4cd0a0