Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> And I suppose this makes some uses of reflect Call a tiny bit faster, avoiding a call at least (https://golang.org/issue/7818). Probably lost in the noise of all the allocations, though.

Yeah, probably lost in the noise, but it might. This might well be the fastest reflect call code path, actually, since the call frame is constructed statically by the compiler and it can always use a frame from a pool. In a sense, this is the "pre-validated" call path you're imagining in #7818, since the validation is done by Value.Interface().

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>