Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Thanks, Austin.
> 
> > Correctness-wise, this seems fine, but I think this may be conflating unrolling the loop with avoiding the overhead of writebarrierptr_prewrite1.
> 
> I had it straight in my head, but yes. :)
> 
> 
> > What do you think about instead lifting some of the setup from writebarrierptr_prewrite1 (like switching to the system stack) into bulkBarrierPreWrite and just calling gcmarkwb_m directly from bulkBarrierPreWrite? Then, as a possible second step, unrolling the bulkBarrierPreWrite loop that calls gcmarkwb_m?
> 
> That was my first instinct, but I wasn't sure whether for a very large bulk pre-write this might exceed the latency budget in a way that hopping back and forth off of the system stack might not. Is this concern ill-founded?

bulkBarrierPreWrite is already non-preemptible, so this wouldn't be making it any worse. In fact, it would be kind of nice for it to be *obviously* non-preemptible, rather than subtly non-preemptible like it is now. :)

This isn't great, obviously. But if we wanted to fix this (which we might have to), I think we would need to do it at the typedmemmove and friends level by breaking it up into smaller bulkBarrierPreWrite and memmove segments with a preemption point after each segment.

Patch-set: 2
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>