doc: net/url ParseQuery example improvement
If you look at the existing ParseQuery example (
https://golang.org/pkg/net/url/#ParseQuery ), you'll see a reference to
a function called toJSON(). I believe this does a huge disservice to the
end user, especially those new to Go.

toJSON() is not a built-in
function, it was a function created specifically for this example (
https://github.com/golang/go/blob/master/src/net/url/example_test.go#L84
).

I removed toJSON() and just did a simple fmt.Println(m) which is still
highly readable ( https://play.golang.org/p/ZR3JuqPXzu ).

Thanks. First time committer.

Change-Id: I7ae53eac8d8f72b11545ba7e616cbc696ce4586e