net/url: improve ParseQuery example
If you look at the existing ParseQuery example (
https://golang.org/pkg/net/url/#ParseQuery ), you'll see a reference to
a function called toJSON(). I believe this does a huge disservice to the
end user, especially those new to Go.

toJSON() is not a built-in
function, it was a function created specifically for this example (
https://github.com/golang/go/blob/master/src/net/url/example_test.go#L84
). toJSON() can't be removed because maps aren't ordered without
additional work in Go.

To resolve this issue, I've moved ExampleParseQuery() and toJSON() to
its own file so that when the end user is reading the documentation,
they understand that toJSON() is not a built in function.

Thanks. First time committer.

Change-Id: I7ae53eac8d8f72b11545ba7e616cbc696ce4586e