Update patch set 1
Patch Set 1:

> no bug number and no test updates?

I was hoping the reason for that would become evident upon looking at the diff.

> This change applies a rare form of inlined panic recovery technique

I was told this rare technique was discovered by a gopher in the Himalayan mountains,
where he spent a year in silence just to better understanding the meaning of
panic, recover and nil.

Now that the date is no longer April 1, upon closer inspection, it is not
a good technique, and the gopher should go back to the drawing board.

The CL description said:

> By throwing more recoveries into the code, we reduce the
> chance of an uncaught panic taking down the user program.

This is untrue, mindlessly throwing more recover calls into the code does not help.

According to the Go specification:

> The return value of recover is nil if any of the following conditions holds:
>
> • panic's argument was nil;
> • the goroutine is not panicking;
> • recover was not called directly by a deferred function.

Also, from https://blog.golang.org/defer-panic-and-recover:

> Recover is only useful inside deferred functions. During normal execution,
> a call to recover will return nil and have no other effect.

In all of the places where recover is added in this CL, recover is not being
called directly by a deferred function, therefore its value will always be
nil.

That explains why the code continues to work with all tests passing,
the behavior is not changed at all. There is no improvement to robustness.

However, the change makes the error handling code much less readable,
not idiomatic, and decreases performance. Therefore it should not be merged.
Abandoning.

Patch-set: 1
Reviewer: Gerrit User 6005 <6005@62eb7196-b449-3ce5-99f1-c037f21e1705>