Update patch set 2
Patch Set 2:

The problem was that in the case of extremely large stack frames, we weren't freeing the Progs. The fix is simple--only flush to machine code when the stack frame is reasonable, but always free the Progs and their associated resources.

Patch-set: 2
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>