net/http/fcgi: expose all cgi environment variables using a new serve interface
The current interface can't access all environment variables directly or via
cgi.RequestFromMap, which only reads variables on its "white list" to be
set on the http.Request it returns. If an fcgi variable is not on the "white list"
 - e.g. REMOTE_USER - the old fcgi api has no access to its value.

This adds another server interface which does have access to all environment variables.

Fixes #16546

Change-Id: Ibf933a768b677ece1bb93d7bf99a14cef36ec671