net/http/fcgi: expose cgi env vars in request context
The current interface can't access all environment
variables directly or via cgi.RequestFromMap, which
only reads variables on its "white list" to be set on
the http.Request it returns. If an fcgi variable is
not on the "white list" - e.g. REMOTE_USER - the old
code no access to its value.

This passes all variables in the Request context and
adds a method that parses the env vars and values from
a request.

Fixes #16546

Change-Id: Ibf933a768b677ece1bb93d7bf99a14cef36ec671