encoding/asn1: allow use of ObjectIdentifier with larger ints
The current implementation assumes int is of size 32.  Furthermore, it
only uses a max of 28 bits of that int.  This change makes it so that
an int64 is used to accumulate up to 63 bits.  This value is then cast
to an int.  If the resulting data does not the overflow the int, it is
used as such.  Thus on a 32 bit machine, up to 31 bits may be used to
represent an ObjectIdentifier.  On a 64 bit machine, up to 63 bits may
be used.

Fixes #19933

Change-Id: I521263dd531e541dbfba264173ea4ed8e918a119